# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API for Coins Monitor.",
        "title": "Coins Monitor Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "saenkodmitriiol@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/bitcoin/add": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Добавляет валюту в список мониторинга",
                "parameters": [
                    {
                        "description": "Имя монеты",
                        "name": "coin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bitcoin/available": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Получает список доступных валют",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.AvailableCurrencyMonitoring"
                            }
                        }
                    }
                }
            }
        },
        "/bitcoin/monitoring": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Получает список мониторингуемых валют",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.AvailableCurrencyMonitoring"
                            }
                        }
                    }
                }
            }
        },
        "/bitcoin/price": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Получает курс биткойна по имени валюты и timestamp",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя монеты",
                        "name": "coin",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "момент времени",
                        "name": "timestamp",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "number",
                            "format": "float64"
                        }
                    }
                }
            }
        },
        "/bitcoin/remove": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Удаляет валюту из списка мониторинга",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя монеты",
                        "name": "coin",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/startmonitor": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Стартует мониторинг",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/stopmonitor": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Останавливает мониторинг",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.AvailableCurrencyMonitoring": {
            "type": "object",
            "properties": {
                "Name": {
                    "type": "string"
                },
                "bitcoin_prices_id": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BitcoinPrice"
                    }
                },
                "coin_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "monitoring": {
                    "type": "boolean"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "model.BitcoinPrice": {
            "type": "object",
            "properties": {
                "bitcoin_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    }
}